refactor: offload logic from stage.tsx to separate files
